org.Hs.egACCNUM {org.Hs.eg.db}	R Documentation

Map Entrez Gene identifiers to GenBank Accession Numbers

Description

org.Hs.egACCNUM is an R object that contains mappings between Entrez Gene identifiers and GenBank accession numbers.

Details

This object is a simple mapping of Entrez Gene identifiers https://www.ncbi.nlm.nih.gov/entrez/query.fcgi?db=gene to all possible GenBank accession numbers.

Mappings were based on data provided by: Entrez Gene ftp://ftp.ncbi.nlm.nih.gov/gene/DATA With a date stamp from the source of: 2021-Sep13

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
x <- org.Hs.egACCNUM
# Get the entrez gene identifiers that are mapped to an ACCNUM
mapped_genes <- mappedkeys(x)
# Convert to a list
xx <- as.list(x[mapped_genes])
if(length(xx) > 0) {
  # Get the ACCNUM for the first five genes
  xx[1:5]
  # Get the first one
  xx[[1]]
}
#For the reverse map ACCNUM2EG:
# Convert to a list
xx <- as.list(org.Hs.egACCNUM2EG)
if(length(xx) > 0){
   # Gets the entrez gene identifiers for the first five Entrez Gene IDs
   xx[1:5]
   # Get the first one
   xx[[1]]
}

[Package org.Hs.eg.db version 3.14.0 Index]