org.Hs.egCHRLENGTHS {org.Hs.eg.db}	R Documentation

A named vector for the length of each of the chromosomes

Description

org.Hs.egCHRLENGTHS provides the length measured in base pairs for each of the chromosomes.

Details

This is a named vector with chromosome numbers as the names and the corresponding lengths for chromosomes as the values.

Total lengths of chromosomes were derived by calculating the number of base pairs on the sequence string for each chromosome.

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
tt <- org.Hs.egCHRLENGTHS
# Length of chromosome 1
tt["1"]

[Package org.Hs.eg.db version 3.14.0 Index]