org.Hs.egMAPCOUNTS {org.Hs.eg.db}	R Documentation

Number of mapped keys for the maps in package org.Hs.eg.db

Description

org.Hs.egMAPCOUNTS provides the "map count" (i.e. the count of mapped keys) for each map in package org.Hs.eg.db.

Details

This "map count" information is precalculated and stored in the package annotation DB. This allows some quality control and is used by the checkMAPCOUNTS function defined in AnnotationDbi to compare and validate different methods (like count.mappedkeys(x) or sum(!is.na(as.list(x)))) for getting the "map count" of a given map.

See Also

mappedkeys

count.mappedkeys

checkMAPCOUNTS

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
org.Hs.egMAPCOUNTS
mapnames <- names(org.Hs.egMAPCOUNTS)
org.Hs.egMAPCOUNTS[mapnames[1]]
x <- get(mapnames[1])
sum(!is.na(as.list(x)))
count.mappedkeys(x)   # much faster!

## Check the "map count" of all the maps in package org.Hs.eg.db
checkMAPCOUNTS("org.Hs.eg.db")

[Package org.Hs.eg.db version 3.14.0 Index]