org.Hs.egOMIM {org.Hs.eg.db}	R Documentation

Map between Entrez Gene Identifiers and Mendelian Inheritance in Man (MIM) identifiers

Description

org.Hs.egOMIM is an R object that provides mappings between entrez gene identifiers and OMIM identifiers.

Details

Each entrez gene identifier is mapped to an OMIM identifier. An NA is reported for any entrez gene identifier that cannot be mapped to an OMIM identifier at this time.

OMIM is based upon the book Mendelian Inheritance in Man (V. A. McKusick) and focuses primarily on inherited or heritable genetic diseases. It contains textual information, pictures, and reference information that can be searched using various terms, among which the MIM number is one.

Mappings were based on data provided by: Entrez Gene ftp://ftp.ncbi.nlm.nih.gov/gene/DATA With a date stamp from the source of: 2021-Sep13

References

https://www.ncbi.nlm.nih.gov/entrez/query.fcgi?db=gene https://www.ncbi.nlm.nih.gov/entrez/query.fcgi?db=OMIM

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
x <- org.Hs.egOMIM
# Get the entrez gene identifiers that are mapped to a OMIM ID
mapped_genes <- mappedkeys(x)
# Convert to a list
xx <- as.list(x[mapped_genes])
if(length(xx) > 0) {
  # Get the OMIM for the first five genes
  xx[1:5]
  # Get the first one
  xx[[1]]
}

# For the reverse map:
x <- org.Hs.egOMIM2EG
# Get the OMIM IDs that are mapped to the entrez gene IDs
mapped_OMIMs <- mappedkeys(x)
# Convert to a list
xx <- as.list(x[mapped_OMIMs])
if(length(xx) > 0) {
  # Get the entrez gene ID for the first five genes
  xx[1:5]
  # Get the first one
  xx[[1]]
}

[Package org.Hs.eg.db version 3.14.0 Index]