org.Hs.egORGANISM {org.Hs.eg.db}	R Documentation

The Organism for org.Hs.eg

Description

org.Hs.egORGANISM is an R object that contains a single item: a character string that names the organism for which org.Hs.eg was built.

Details

Although the package name is suggestive of the organism for which it was built, org.Hs.egORGANISM provides a simple way to programmatically extract the organism name.

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
org.Hs.egORGANISM

[Package org.Hs.eg.db version 3.14.0 Index]