org.Hs.eg_dbconn {org.Hs.eg.db}	R Documentation

Collect information about the package annotation DB

Description

Some convenience functions for getting a connection object to (or collecting information about) the package annotation DB.

Usage

  org.Hs.eg_dbconn()
  org.Hs.eg_dbfile()
  org.Hs.eg_dbschema(file="", show.indices=FALSE)
  org.Hs.eg_dbInfo()

Arguments

file
A connection, or a character string naming the file to print to (see the file argument of the cat function for the details).

show.indices
The CREATE INDEX statements are not shown by default. Use show.indices=TRUE to get them.

Details

org.Hs.eg_dbconn returns a connection object to the package annotation DB. IMPORTANT: Don't call dbDisconnect on the connection object returned by org.Hs.eg_dbconn or you will break all the AnnDbObj objects defined in this package!

org.Hs.eg_dbfile returns the path (character string) to the package annotation DB (this is an SQLite file).

org.Hs.eg_dbschema prints the schema definition of the package annotation DB.

org.Hs.eg_dbInfo prints other information about the package annotation DB.

Value

org.Hs.eg_dbconn: a DBIConnection object representing an open connection to the package annotation DB.

org.Hs.eg_dbfile: a character string with the path to the package annotation DB.

org.Hs.eg_dbschema: none (invisible NULL).

org.Hs.eg_dbInfo: none (invisible NULL).

See Also

dbGetQuery, dbConnect, dbconn, dbfile, dbschema, dbInfo

Examples

  library(DBI)
  ## Count the number of rows in the "genes" table:
  dbGetQuery(org.Hs.eg_dbconn(), "SELECT COUNT(*) FROM genes")

  org.Hs.eg_dbschema()

  org.Hs.eg_dbInfo()

[Package org.Hs.eg.db version 3.14.0 Index]